Added correctness predicates
Predicate for a forest being correct (does not include ordering of the
keys), and a predicate for a cursor being correct based off of it.

Also updated task list in README

The cursor_correct predicate does include cursor splits being nested.
That is, if (f1,f2)::(f1',f2')::tl is a valid cursor, then (zip f1 f2)
must be the sub-forest of the first tree of f2'.